soc.csr_types: re-home PokeReg as PulseReg
